Refactor: getTicketCustomObject 위치 변경
