+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/ticket/custom/query/controller/TicketCustomQueryController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ticketaka.mtvs3_final_backend._core.utils.ApiUtils;
-import ticketaka.mtvs3_final_backend.ticketing.ticket.custom.query.dto.TicketCustomQueryResponseDTO;
 import ticketaka.mtvs3_final_backend.ticketing.ticket.custom.query.service.TicketCustomQueryService;
 
 import static ticketaka.mtvs3_final_backend._core.utils.SecurityUtils.getCurrentMemberId;
@@ -18,21 +17,4 @@ import static ticketaka.mtvs3_final_backend._core.utils.SecurityUtils.getCurrent
 @RestController
 @RequestMapping("/api/member/tickets")
 public class TicketCustomQueryController {
-
-    private final TicketCustomQueryService ticketCustomQueryService;
-
-    /*
-        티켓 커스텀 제작 입장 - 스티커 조회
-     */
-    @GetMapping("/{ticketId}/custom")
-    public ResponseEntity<?> getTicketCustomObject(@PathVariable("ticketId") Long ticketId) {
-
-        log.info("getTicketCustomInfo Request");
-
-        TicketCustomQueryResponseDTO.getTicketCustomObjectDTO responseDTO = ticketCustomQueryService.getTicketCustomObject(getCurrentMemberId(), ticketId);
-
-        log.info("getTicketCustomInfo Response : {}", responseDTO);
-
-        return ResponseEntity.ok().body(ApiUtils.success(responseDTO));
-    }
 }
