+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/ticket/query/dto/TicketQueryResponseDTO.java
@@ -1,7 +1,5 @@
 package ticketaka.mtvs3_final_backend.ticketing.ticket.query.dto;
 
-import ticketaka.mtvs3_final_backend.ticketing.ticket.custom.query.dto.TicketCustomQueryResponseDTO;
-
 import java.util.List;
 
 public class TicketQueryResponseDTO {
@@ -29,6 +27,22 @@ public class TicketQueryResponseDTO {
             int month,
             int day,
             String time
+    ) {
+    }
+
+    /*
+        티켓 커스텀 입장
+     */
+    public record getTicketCustomObjectDTO(
+            int dailyBackgroundRefreshCount,
+            List<stickerDTO> stickerDTOList
+    ) {
+    }
+
+    // Sticker
+    public record stickerDTO(
+            int stickerId,
+            byte[] stickerImage
     ) {
     }
 }
